[Refactor] #57 - UserInfo 관련 네트워크 코드 분리 및 개선
