+++ Memento-iOS/Memento-iOS/Source/Network/Onboarding/APIService/UserInfoAPIService.swift
@@ -11,7 +11,7 @@ import Moya
 // MARK: - UserInfoAPIServiceProtocol
 
 protocol UserInfoAPIServiceProtocol {
-    func updateUserInfo(request: UserInfoRequest, completion: @escaping (NetworkResult<EmptyDTO>) -> Void)
+    func updateUserInfo(body: UserInfoRequest, completion: @escaping (NetworkResult<EmptyDTO>) -> Void)
 }
 
 // MARK: - UserInfoAPIService
@@ -23,9 +23,8 @@ final class UserInfoAPIService: BaseAPIService, UserInfoAPIServiceProtocol {
         plugins: [MoyaPlugin.shared]
     )
     
-    /// 사용자 정보 업데이트 API 호출
-    func updateUserInfo(request: UserInfoRequest, completion: @escaping (NetworkResult<EmptyDTO>) -> Void) {
-        provider.request(UserInfoTargetType.updateUserInfo(request: request)) { [weak self] result in
+    func updateUserInfo(body: UserInfoRequest, completion: @escaping (NetworkResult<EmptyDTO>) -> Void) {
+        provider.request(UserInfoTargetType.updateUserInfo(body: body)) { [weak self] result in
             guard let self = self else { return }
 
             switch result {
