+++ Memento-iOS/Memento-iOS/Source/Network/Setting/DTO/Response/UserUptimeResponse.swift
